+++ admin/formuledevoyage_extrafields.php
@@ -92,7 +92,7 @@ require DOL_DOCUMENT_ROOT.'/core/actions_extrafields.inc.php';
  * View
  */
 
-$textobject = $langs->transnoentitiesnoconv("FormuledeVoyage");
+$textobject = $langs->transnoentitiesnoconv("FormuleDeVoyage");
 
 $help_url = '';
 $page_name = "CliEnjoyHolidaysSetup";