+++ class/actions_clienjoyholidays.class.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * Class ActionsSupplierorderfromorder
+ *
+ * Hook actions
+ */
+
+class ActionsClienjoyHolidays
+{
+	function addMoreActionsButtons($parameters, &$object, &$action, $hookmanager){
+		global $langs;
+		$url = dol_buildpath('/clienjoyholidays/formuledevoyage_card.php', 1).'?origin_id='.$object->id.'&origin='.$object->element.'&action=create';
+		print dolGetButtonAction('', $langs->trans('createFormuleDeVoyageFromPropal'), 'create', $url, '');
+
+		return 0;
+	}
+}